combined PipelineManager.hpp and DescriptorSetManager.hpp.
